Extracting parsers to own classes
