+++ greenmail-core/src/main/java/com/icegreen/greenmail/imap/commands/ListCommand.java
@@ -10,6 +10,7 @@ import com.icegreen.greenmail.imap.ImapRequestLineReader;
 import com.icegreen.greenmail.imap.ImapResponse;
 import com.icegreen.greenmail.imap.ImapSession;
 import com.icegreen.greenmail.imap.ProtocolException;
+import com.icegreen.greenmail.imap.commands.parsers.ListCommandParser;
 import com.icegreen.greenmail.store.FolderException;
 import com.icegreen.greenmail.store.MailFolder;
 import com.sun.mail.imap.protocol.BASE64MailboxDecoder; // NOSONAR
@@ -161,38 +162,6 @@ public class ListCommand extends AuthenticatedStateCommand {
 
         buffer.insert(0, referenceName);
         return buffer.toString();
-    }
-
-    private static class ListCommandParser extends CommandParser {
-        /**
-         * Reads an argument of type "list_mailbox" from the request, which is
-         * the second argument for a LIST or LSUB command. Valid values are a "string"
-         * argument, an "atom" with wildcard characters.
-         *
-         * @return An argument of type "list_mailbox"
-         */
-        public String listMailbox(ImapRequestLineReader request) throws ProtocolException {
-            char next = request.nextWordChar();
-            String name;
-            switch (next) {
-                case '"':
-                    name = consumeQuoted(request);
-                    break;
-                case '{':
-                    name = consumeLiteral(request);
-                    break;
-                default:
-                    name = consumeWord(request, new ListCharValidator());
-            }
-            return BASE64MailboxDecoder.decode(name);
-        }
-
-        private class ListCharValidator extends AtomCharValidator {
-            @Override
-            public boolean isValid(char chr) {
-                return isListWildcard(chr) || super.isValid(chr);
-            }
-        }
     }
 }
