+++ greenmail-core/src/main/java/com/icegreen/greenmail/imap/commands/StatusCommand.java
@@ -10,6 +10,8 @@ import com.icegreen.greenmail.imap.ImapRequestLineReader;
 import com.icegreen.greenmail.imap.ImapResponse;
 import com.icegreen.greenmail.imap.ImapSession;
 import com.icegreen.greenmail.imap.ProtocolException;
+import com.icegreen.greenmail.imap.commands.parsers.StatusCommandParser;
+import com.icegreen.greenmail.imap.commands.parsers.status.StatusDataItems;
 import com.icegreen.greenmail.store.FolderException;
 import com.icegreen.greenmail.store.MailFolder;
 import com.sun.mail.imap.protocol.BASE64MailboxEncoder; // NOSONAR
@@ -20,15 +22,15 @@ import com.sun.mail.imap.protocol.BASE64MailboxEncoder; // NOSONAR
  * @author Darrell DeBoer <darrell@apache.org>
  * @version $Revision: 109034 $
  */
-class StatusCommand extends AuthenticatedStateCommand {
+public class StatusCommand extends AuthenticatedStateCommand {
     public static final String NAME = "STATUS";
     public static final String ARGS = "<mailbox> ( <status-data-item>+ )";
 
-    private static final String MESSAGES = "MESSAGES";
-    private static final String RECENT = "RECENT";
-    private static final String UIDNEXT = "UIDNEXT";
-    private static final String UIDVALIDITY = "UIDVALIDITY";
-    private static final String UNSEEN = "UNSEEN";
+    public static final String MESSAGES = "MESSAGES";
+    public static final String RECENT = "RECENT";
+    public static final String UIDNEXT = "UIDNEXT";
+    public static final String UIDVALIDITY = "UIDVALIDITY";
+    public static final String UNSEEN = "UNSEEN";
 
     private StatusCommandParser statusParser = new StatusCommandParser();
 
@@ -61,35 +63,35 @@ class StatusCommand extends AuthenticatedStateCommand {
         buffer.append(SP);
         buffer.append('(');
 
-        if (statusDataItems.messages) {
+        if (statusDataItems.isMessages()) {
             buffer.append(MESSAGES);
             buffer.append(SP);
             buffer.append(folder.getMessageCount());
             buffer.append(SP);
         }
 
-        if (statusDataItems.recent) {
+        if (statusDataItems.isRecent()) {
             buffer.append(RECENT);
             buffer.append(SP);
             buffer.append(folder.getRecentCount(false));
             buffer.append(SP);
         }
 
-        if (statusDataItems.uidNext) {
+        if (statusDataItems.isUidNext()) {
             buffer.append(UIDNEXT);
             buffer.append(SP);
             buffer.append(folder.getUidNext());
             buffer.append(SP);
         }
 
-        if (statusDataItems.uidValidity) {
+        if (statusDataItems.isUidValidity()) {
             buffer.append(UIDVALIDITY);
             buffer.append(SP);
             buffer.append(folder.getUidValidity());
             buffer.append(SP);
         }
 
-        if (statusDataItems.unseen) {
+        if (statusDataItems.isUnseen()) {
             buffer.append(UNSEEN);
             buffer.append(SP);
             buffer.append(folder.getUnseenCount());
@@ -103,53 +105,6 @@ class StatusCommand extends AuthenticatedStateCommand {
 
         session.unsolicitedResponses(response);
         response.commandComplete(this);
-    }
-
-    private static class StatusCommandParser extends CommandParser {
-        StatusDataItems statusDataItems(ImapRequestLineReader request)
-                throws ProtocolException {
-            StatusDataItems items = new StatusDataItems();
-
-            request.nextWordChar();
-            consumeChar(request, '(');
-            CharacterValidator validator = new NoopCharValidator();
-            String nextWord = consumeWord(request, validator);
-            while (!nextWord.endsWith(")")) {
-                addItem(nextWord, items);
-                nextWord = consumeWord(request, validator);
-            }
-            // Got the closing ")", may be attached to a word.
-            if (nextWord.length() > 1) {
-                addItem(nextWord.substring(0, nextWord.length() - 1), items);
-            }
-
-            return items;
-        }
-
-        private void addItem(String nextWord, StatusDataItems items)
-                throws ProtocolException {
-            if (nextWord.equals(MESSAGES)) {
-                items.messages = true;
-            } else if (nextWord.equals(RECENT)) {
-                items.recent = true;
-            } else if (nextWord.equals(UIDNEXT)) {
-                items.uidNext = true;
-            } else if (nextWord.equals(UIDVALIDITY)) {
-                items.uidValidity = true;
-            } else if (nextWord.equals(UNSEEN)) {
-                items.unseen = true;
-            } else {
-                throw new ProtocolException("Unknown status item: '" + nextWord + '\'');
-            }
-        }
-    }
-
-    private static class StatusDataItems {
-        boolean messages;
-        boolean recent;
-        boolean uidNext;
-        boolean uidValidity;
-        boolean unseen;
     }
 }
