+++ greenmail-core/src/main/java/com/icegreen/greenmail/imap/commands/parsers/SearchCommandParser.java
@@ -4,10 +4,12 @@
  * This file has been modified by the copyright holder.
  * Original file can be found at http://james.apache.org
  */
-package com.icegreen.greenmail.imap.commands;
+package com.icegreen.greenmail.imap.commands.parsers;
 
 import com.icegreen.greenmail.imap.ImapRequestLineReader;
 import com.icegreen.greenmail.imap.ProtocolException;
+import com.icegreen.greenmail.imap.commands.SearchKey;
+import com.icegreen.greenmail.imap.commands.parsers.search.SearchTermBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,7 +42,7 @@ public class SearchCommandParser extends CommandParser {
         SearchTerm resultTerm = null;
         SearchTermBuilder b = null;
         SearchKey key = null;
-	boolean orKey = false;
+        boolean orKey = false;
         boolean negated = false;
         // Dummy implementation
         // Consume to the end of the line.
